:sparkles: Logs view in admin page

+++ src/routes/components/email.route.php
@@ -30,11 +30,15 @@ Route::group(['prefix' => 'logs/email'], function () {
     Route::post('/remove', 'Backend\EmailLogController@remove')->name('admin.logs.email.remove');
 });
 
-Route::group(['prefix' => 'logs/error'], function () {
-    Route::get('/', 'Backend\LogViewerController@index')->name('admin.logs.error');
-    Route::get('/get-data', 'Backend\LogViewerController@listLogs')->name('admin.logs.error.get-logs');
-    Route::delete('/delete', 'Backend\LogViewerController@delete')->name('admin.logs.delete');
+if (config('juzaweb.logs_viewer')) {
+    Route::group(['prefix' => 'logs/error'], function () {
+        Route::get('/', 'Backend\LogViewerController@index')->name('admin.logs.error');
+        Route::get('/get-data', 'Backend\LogViewerController@listLogs')->name('admin.logs.error.get-logs');
+        Route::delete('/delete', 'Backend\LogViewerController@delete')->name('admin.logs.delete');
 
-    Route::get('/{date}', 'Backend\LogViewerController@show')->name('admin.logs.error.date');
-    Route::get('/{date}/get-data', 'Backend\LogViewerController@listLogsDate')->name('admin.logs.error.get-logs-date');
-});
+        Route::get('/{date}', 'Backend\LogViewerController@show')->name('admin.logs.error.date');
+        Route::get('/{date}/get-data', 'Backend\LogViewerController@listLogsDate')->name('admin.logs.error.get-logs-date');
+    });
+}
+
+
